refactor(gui): typo correction
